scripts/requirements.txt
garminconnect>=0.2.0
python-dotenv>=1.0.0

